+++ inst/include/gdalwarpgeneral/gdalwarpgeneral.h
@@ -192,7 +192,9 @@ inline List gdal_warp_general(CharacterVector dsn,
     papszArg = CSLAddString(papszArg, resample[0]);
   }
   for (int gwopt = 0; gwopt < options.length(); gwopt++) {
-    papszArg = CSLAddString(papszArg, options[gwopt]);
+    if (!options[gwopt].empty()) {
+     papszArg = CSLAddString(papszArg, options[gwopt]);
+    }
   }