+++ NEWS.md
@@ -1,5 +1,7 @@
 # vapour 0.10.0
 
+* Fixed putting bad options in (empty strings). 
+
 * Fixed type problem in internal get projection strings. 
 
 * Urgent release to fix segfault on CRAN, instances of int instead of IntegerVector for R devel.